+++ at_contact/lib/src/config/app_constants.dart
@@ -1,3 +1,4 @@
+// ignore_for_file: non_constant_identifier_names
 class AppConstants {
   static final String CONTACT_KEY_PREFIX = 'atconnections';
   static final String CONTACT_KEY_SUFFIX = 'contacts';