+++ at_contact/test/at_contact_tests.dart
@@ -10,7 +10,8 @@ Future<void> main() async {
   late AtContact contact;
   var atSign = '@vinod';
   try {
-    await AtClientManager.getInstance().setCurrentAtSign(atSign, 'me', TestUtil.getPreferenceLocal());
+    await AtClientManager.getInstance()
+        .setCurrentAtSign(atSign, 'me', TestUtil.getPreferenceLocal());
     atContact = await AtContactsImpl.getInstance('@colin');
     // set contact details
     contact = AtContact(