+++ at_lookup/lib/src/exception/at_lookup_exception.dart
@@ -4,7 +4,13 @@ import 'package:at_commons/at_commons.dart';
 class AtLookUpException implements Exception {
   String? errorCode;
   String? errorMessage;
+
   AtLookUpException(this.errorCode, this.errorMessage);
+
+  @override
+  String toString() {
+    return 'ErrorCode: $errorCode - Exception: $errorMessage';
+  }
 }
 
 /// AtLookUpExceptionUtil to get ErrorCode and ErrorDescription