+++ at_contact/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.0.2
+- uptake AtKey validations from at_commons package. 
+## 3.0.1
+- at_client version upgrade
 ## 3.0.0
 - RSDK uptake
 ## 2.0.3