+++ at_lookup/CHANGELOG.md
@@ -1,5 +1,7 @@
-## 3.0.4
+## 3.0.5
 - at_commons and at_utils version change
+## 3.0.4
+- at_utils and at_commons version change for AtKey validations. 
 ## 3.0.3
 - Reduce wait time on address lookup to root server 
 ## 3.0.2